--- app/src/main/java/com/walts/ideas/LocationResult.java
@@ -1,7 +1,0 @@
-package com.walts.ideas;
-
-import android.location.Location;
-
-public abstract class LocationResult {
-    public abstract void gotLocation(Location location);
-}
