--- app/src/main/java/com/walts/ideas/MyLocation.java
@@ -1,112 +1,0 @@
-package com.walts.ideas;
-
-import java.util.Timer;
-import java.util.TimerTask;
-import android.content.Context;
-import android.location.Location;
-import android.location.LocationListener;
-import android.location.LocationManager;
-import android.os.Bundle;
-
-//http://stackoverflow.com/questions/3145089/what-is-the-simplest-and-most-robust-way-to-get-the-users-current-location-in-a/3145655#3145655
-public class MyLocation {
-
-    private static final String TAG = "MyLocation";
-
-    private Timer timer = new Timer();
-
-    private LocationManager locationManager;
-    private LocationResult locationResult;
-
-    private boolean gpsEnabled = false;
-    private boolean networkEnabled = false;
-
-    public MyLocation(Context context) {
-        locationManager = (LocationManager) context.getSystemService(Context.LOCATION_SERVICE);
-    }
-
-    public boolean canGetLocation(LocationResult result) {
-        locationResult = result;
-        gpsEnabled = locationManager.isProviderEnabled(LocationManager.GPS_PROVIDER);
-        networkEnabled = locationManager.isProviderEnabled(LocationManager.NETWORK_PROVIDER);
-        if (!gpsEnabled && !networkEnabled) {
-            return false;
-        }
-        if (gpsEnabled) {
-            locationManager.requestLocationUpdates(LocationManager.GPS_PROVIDER, 0, 0, locationListenerGps);
-        }
-        if (networkEnabled) {
-            locationManager.requestLocationUpdates(LocationManager.NETWORK_PROVIDER, 0, 0, locationListenerNetwork);
-        }
-        timer.schedule(new GetLastLocation(), 5000);
-        return true;
-    }
-
-    //you may need to call this onPause() in activity
-    public void cancel() {
-        timer.cancel();
-        locationManager.removeUpdates(locationListenerGps);
-        locationManager.removeUpdates(locationListenerNetwork);
-    }
-
-    private LocationListener locationListenerGps = new LocationListener() {
-        public void onLocationChanged(Location location) {
-            timer.cancel();
-            locationResult.gotLocation(location);
-            locationManager.removeUpdates(this);
-            locationManager.removeUpdates(locationListenerNetwork);
-        }
-        public void onProviderDisabled(String provider) {}
-        public void onProviderEnabled(String provider) {}
-        public void onStatusChanged(String provider, int status, Bundle extras) {}
-    };
-
-    private LocationListener locationListenerNetwork = new LocationListener() {
-        public void onLocationChanged(Location location) {
-            timer.cancel();
-            locationResult.gotLocation(location);
-            locationManager.removeUpdates(this);
-            locationManager.removeUpdates(locationListenerGps);
-        }
-        public void onProviderDisabled(String provider) {}
-        public void onProviderEnabled(String provider) {}
-        public void onStatusChanged(String provider, int status, Bundle extras) {}
-    };
-
-    private class GetLastLocation extends TimerTask {
-
-        private static final String TAG = "GetLastLocation";
-
-        @Override
-        public void run() {
-            locationManager.removeUpdates(locationListenerGps);
-            locationManager.removeUpdates(locationListenerNetwork);
-            Location gpsLocation = null;
-            Location networkLocation = null;
-            if (gpsEnabled) {
-                gpsLocation = locationManager.getLastKnownLocation(LocationManager.GPS_PROVIDER);
-            }
-            if (networkEnabled) {
-                networkLocation = locationManager.getLastKnownLocation(LocationManager.NETWORK_PROVIDER);
-            }
-            if (gpsLocation != null && networkLocation != null) {
-                if (gpsLocation.getTime() > networkLocation.getTime()) {
-                    locationResult.gotLocation(gpsLocation);
-                } else {
-                    locationResult.gotLocation(networkLocation);
-                }
-                return;
-            }
-            if (gpsLocation != null) {
-                locationResult.gotLocation(gpsLocation);
-                return;
-            }
-            if (networkLocation != null) {
-                locationResult.gotLocation(networkLocation);
-                return;
-            }
-            locationResult.gotLocation(null);
-        }
-    }
-
-}
